backend/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
google-generativeai
google-genai
pydantic
